Phone Inclinometer: FIX app not responsive
The app was not responsive cause we were sending a http request with
inclinometer data every 0.5 seconds, the code was running but the app
was getting unresponsive.

This commit adds
- max sampling: instead of sending every value to server, we take the
  max of every 20 values. We don't need that level of accuracy
- async task: instead of executing the http request on main thread we
  execute it in the background

TODO:
- open the Logcat and filter logs with "androidAccellerometer", check
  how often we send the logs to the server: we still send them too
often! set the parmatere so that we send them every 2 minutes
- instead of sending x,y,z we need to send inclination in degrees. I
  need to know how many degrees is the phone inclined

Test Plan:
running app on phone
Logging on android studio working (screenshot 1)
data coming to influx (screenshot 2)