[sensor] fix esp32 bug, now motors working and sensor working
looks like we didn't test esp32 on this branch, there was a bug
introduced earlier where we were trying to send android sensor data in
the ESP32 data. If we don't test with esp on there is a whole part of
the app that won't be running so we realy need to do it.